pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vrcpost-poster"
version = "0.1.0"
description = "Scheduled posting and story auto-update tool for VRCPost"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Meta-Develop", email = "rikuuyoo@gmail.com"}]
dependencies = [
    "playwright>=1.40.0",
    "schedule>=1.2.0",
    "Pillow>=10.0.0",
    "customtkinter>=5.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "cryptography>=41.0.0",
    "apscheduler>=3.10.0",
    "flask>=3.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
vrcpost-poster = "src.gui.main_window:main"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
